Refactor `ModelInspector`
