+++ src/routes/_surfaces/inspector/index.svelte
@@ -1,17 +1,15 @@
 <script lang="ts">
-  import { getContext } from "svelte";
-  import ModelInspector from "./Model.svelte";
-
-  import type { ApplicationStore } from "$lib/application-state-stores/application-stores";
-
-  import { drag } from "$lib/drag";
   import { EntityType } from "$common/data-modeler-state-service/entity-state-service/EntityStateService";
+  import type { ApplicationStore } from "$lib/application-state-stores/application-store";
   import {
     inspectorVisibilityTween,
     inspectorVisible,
     layout,
   } from "$lib/application-state-stores/layout-store";
   import Portal from "$lib/components/Portal.svelte";
+  import { drag } from "$lib/drag";
+  import { getContext } from "svelte";
+  import ModelInspector from "./ModelInspector.svelte";
 
   const store = getContext("rill:app:store") as ApplicationStore;
 </script>
